docs/ and code/
I kept getting a syntax error message when cloning in github. I couldn't even start the github parts of the assignment. Worked on this for about 2 hours and finally threw in the towel. I thought I followed the concepts well presented yesterday morning, however this still seems to be a challenge for me. I will reach out to Ivan to see if he can show me how to perform those tasks so I will have notes for future use.