Update 05-This-Is-Only-A-Test-Joe Cursi
After our train hit the minivan, I picked my fine apple product and attempted to start this single task. So from 5:33 I followed the directions as stated in our requirements and  “Go to Ruby Koans and download the koans. Follow the instructions the site gives you, and work through the end of the about_symbols.rb file.”  So I go to the site and run the download. Unlike a windows product that asks if I would like to install the file, it just appears in my ‘Download’ folder. So I open the file and a bunch of text files. The first appropriate direction that I get is that, “You can run the tests by calling the path_to_enlightenment.rb file.” So not knowing what this means I proceed to attempt to paste the contents of the text file into my terminal and hit enter, hoping this will load something that I will be able to run later. Nothing happens. So what I do next is follow the next set of directions, being; “In your terminal, while in the ruby_koans directory, type:[ ruby_koans ] $ ruby path_to_enlightenment.rb”. So after this I begin searching for the ruby.koans directory, thinking that is was installed. At 6:55 the train started moving again so I decided that an hour and 22 minutes of this was attempt enough.